# docs/swagger.yaml
definitions:
...
  ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  TokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      mfa_token:
        type: string
      requires_password_change:
        type: boolean
      requires_totp:
        type: boolean
      token_type:
        type: string
    type: object
...
  mfa.SetupResult:
    properties:
      qr:
        type: string
      secret:
        type: string
      uri:
        type: string
    type: object
  mfa.Status:
    properties:
      enabled:
        type: boolean
      local:
        type: boolean
      recovery_remaining:
        type: integer
    type: object
...
  v1_users.RecoveryResponse:
    properties:
      access_token:
        type: string
      recovery_codes:
        items:
          type: string
        type: array
    type: object
  v1_users.TOTPRequest:
    properties:
      code:
        type: string
      mfa_token:
        type: string
      new_password:
        type: string
      password:
        type: string
    type: object
info:
  contact: {}
  description: API for interacting with Marmot
  license:
    name: MIT
    url: https://opensource.org/license/MIT
  title: Marmot API
  version: "0.1"
paths:
...
  /api/v1/users/login/password:
    post:
      consumes:
      - application/json
      operationId: changeLoginPassword
      parameters:
      - description: Verification request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1_users.TOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Change a mandatory password before the second factor
      tags:
      - users
  /api/v1/users/login/totp:
    post:
      consumes:
      - application/json
      operationId: verifyTOTPLogin
      parameters:
      - description: Verification request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1_users.TOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Complete local login using a second factor
      tags:
      - users
...
  /api/v1/users/totp:
    delete:
      consumes:
      - application/json
      operationId: disableTOTP
      parameters:
      - description: Verification request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1_users.TOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1_users.RecoveryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Disable TOTP with password and second factor
      tags:
      - users
    get:
      consumes:
      - application/json
      operationId: totpStatus
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mfa.Status'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get local two-factor status
      tags:
      - users
  /api/v1/users/totp/confirm:
    post:
      consumes:
      - application/json
      operationId: confirmTOTP
      parameters:
      - description: Verification request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1_users.TOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1_users.RecoveryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Confirm TOTP and return recovery codes once
      tags:
      - users
  /api/v1/users/totp/recovery:
    post:
      consumes:
      - application/json
      operationId: regenerateRecovery
      parameters:
      - description: Verification request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1_users.TOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1_users.RecoveryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Replace one-use recovery codes
      tags:
      - users
  /api/v1/users/totp/reset/{id}:
    delete:
      consumes:
      - application/json
      operationId: resetTOTP
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Administratively reset a local second factor
      tags:
      - users
  /api/v1/users/totp/setup:
    post:
      consumes:
      - application/json
      operationId: setupTOTP
      parameters:
      - description: Verification request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1_users.TOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mfa.SetupResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Prepare an encrypted TOTP enrollment
      tags:
      - users
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"